Add legend for Discharge Date
Per feedback from David Ondich:
```
navigating strictly using tab key, I only received that you wanted the month, but had to use my arrow keys to determine that you were asking about date of discharge; The rest of the edit boxes and radio buttons were totally navigable.
```

So the only difference between that and the date of birth input is the absence of an explicit `<legend>` tag. Long-term we could apply a supplemental class to the `<legend>` tag when it's used more inline as in this context, but that's some trivial CSS messery.